Use SSE4_CRC32 v4 to support node.js v4 and above
The use of the fast-crc32c library doesn't allow the hash-stream-validation
module to work on node.js v4+ due to its dependency on an older version that
does not use Nan v2 (which is required for node.js add-ons to work on node.js v4
and above).

This commit brings node.js v4 support to the hash-stream-validation libary.